Working build of the game
It has none of what I initially wanted but I wanted a lot and it would be really hard to make.
The description of the game is nowhere near what it is now because I didn't realize making an ace attorney
game with raw source code would be hard. The issue is mainly that I have no way of measuring the player's progress due to how disconnected everything is.